Moved dependencies away from ReadModel. ReadModel contains IEventStream, implemented in EventReader, and dependency injected by Bootstrap.
