+++ Bootstrap/Bootstrap.csproj
@@ -10,6 +10,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\EventReader\EventReader.csproj" />
     <ProjectReference Include="..\ReadModel\ReadModel.csproj" />
   </ItemGroup>
