+++ EventReader/EventReader.csproj
@@ -9,4 +9,8 @@
     <PackageReference Include="System.Data.SqlClient" Version="4.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ReadModel\ReadModel.csproj" />
+  </ItemGroup>
+
 </Project>
