+++ EventReader/ISerialize.cs
@@ -1,0 +1,9 @@
+﻿using ReadModel.Events;
+
+namespace EventReader
+{
+    public interface ISerialize : IEventStream
+    {
+        IEvent DeserializeEntry(EventEntry entry);
+    }
+}
