+++ ReadModel/Events/IEventStream.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ReadModel.Events
+{
+    public interface IEventStream
+    {
+        IEnumerable<IEvent> Get(EventType eventType);
+    }
+}
