+++ ReadModel/ModelGenerator.cs
@@ -1,5 +1,5 @@
-﻿using EventReader.Event;
-using EventReader.Read;
+﻿using ReadModel.Events;
+using ReadModel.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,24 +8,9 @@ namespace ReadModel
 {
     public class ModelGenerator
     {
-        EventStream Stream;
+        IEventStream Stream;
 
-        // TODO: might still use CUSTOMER_DETAILS_QUERY below...
-        // IDEA: Rather than build a separate stream of CustomerCreated and RepaymentTaken events...
-            // Iterate each RepaymentTaken event
-                // If Customer doesn't exist in CustomerModel
-                    // Call Stream.Get() to retrieve single CustomerCreated event, where PaymentEvent.CustomerId == AggregateId
-                    // Create new Customer and add to CustomerModel
-                // Add payment to Customer
-        const string CUSTOMER_DETAILS_QUERY =
-            "select * " +
-            "from MessageHub.Message as t0 " +
-            "join MessageHub.MessageContent as t1 " +
-            "on t0.SequenceId = t1.SequenceId " +
-            "where t0.MessageTypeId = 1 and t0.AggregateTypeId = 11 and " +
-            "t0.AggregateId = @customerId ";
-
-        public ModelGenerator(EventStream stream)
+        public ModelGenerator(IEventStream stream)
         {
             Stream = stream;
         }
@@ -33,8 +18,8 @@ namespace ReadModel
         public Dictionary<Guid, Customer> GetCustomerModel()
         {
             var customerModel = new Dictionary<Guid, Customer>();
-            var customerEvents = Stream.Get(EventType.CustomerCreated).Take(50000);
-            var paymentEvents = Stream.Get(EventType.RepaymentTaken).Take(50000);
+            var customerEvents = Stream.Get(EventType.CustomerCreated).Take(10000);
+            var paymentEvents = Stream.Get(EventType.RepaymentTaken).Take(10000);
 
             foreach (CustomerCreatedEvent cce in customerEvents)
             {
