+++ ReadModel/Models/MonthYearPayment.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ReadModel
+﻿namespace ReadModel.Models
 {
     public class MonthYearPayment
     {
