+++ ReadModel/ReadModel.csproj
@@ -4,8 +4,4 @@
     <TargetFramework>netcoreapp1.1</TargetFramework>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\EventReader\EventReader.csproj" />
-  </ItemGroup>
-
 </Project>
